Remove DataMapper-specific code from tests
I've checked the DM-specific code and most of it was possible to remove without
any problems. There's however an issue that sometimes the class needs to be
finalized after manipulating its structure. Therefore a new method was added to
internal helpers "udpate_model", which will just execute block for other ORMs,
but in DM will also finalize the class