Replace `migrations` generator option with `only-submodules`
Gnerator has an option `--migrations` that actually does not do what its name
says. It also changes user model and initializer. The real change it does is
that if it's provided, no core files are copied (user model is not generated,
core migrations are not copied).

Therefore the proper name for this option is `--only-submodules`.
THe previous option was deprecated and should be removed in v0.9.0.