Implement new logout_all_sessions API method, which forces clearing the remember me token
Normally logout will not invalidate the remember me token if
remember_me_token_persist_globally is set to true, since other browser
may still be relying on that session.

We now expose a new logout_all_sessions method which forces this token
value to be cleared, effectively destroying remember me for that user on
all browsers.

Note that the implementation is a bit wonky due to the coupling with the
callback queues we have. Must of this messiness can be removed if we
decide to just expose force_forget_me! instead of a new wrapper
logout_all_sessions.