fixes bug in linkedin where id returns blank
LinkedIn doesn't automatically return an id in v1/people/~ 

So, if you don't specify `id` in the call, the `uid` in `Authentications` will be blank, and only the first user will be created 'successfully'. This change adds `id` to the call by default so it doesn't need specified manually by the developer in the initializer's `config.linkedin.user_info_fields`.